Adds keys and locks
Adds locks as a steel sheet recipe that can be used on any /obj/structure/simple_door to create locked doors.

Uses the var/uid to generate new uid for every lock when made.

Locks can be applied to doors that will then drop a key. Unlocked doors with a lock can have a lock applied to them, setting the lock to be the same. Keys drop after you apply them to a door, meaning there will be a number of keys equal to the number of applied locks.

This is probably not the optimal way to go about things and is a pile of if/else statements, but here we go.