Adds locks and keys (fukk me)
Made the locking proc actually function on /door. There's probably some spaghetti rendundancy that needs fixing in terms of operating doors, (open), bumpopen, and other shenanigans but I'm not doing it right now.

Added descriptions of key and lock under New() to ensure that people can check what number the keys and locks are set to.

With steel sheets in hand:
Manufacture a lock (or multiple).
Manufacture a key (or multiple).
Use the keys on the locks to set them to be the same as that lock.
Use the locks on other locks to set them to be the same.
Apply the locks to doors.
While a door is unlocked, apply a lock to set it to be the same.
If you have a key, apply a single rod to the key to manufacture an identical key.

Things that still need doing: An overlay sprite for doors with a non-ugly lock, better lock, in-hand sprites for both lock and key.

Special thanks to Rhicora and d20 folks who helped me out with this.

Added bonus: Comes with a map fix that replaces a space square with desert (Mr. Moose pointed this out).