Fixes nonexistant icon of tiled desert
Refactoring and some cleanup as well.

Problems: Automatically reconfigures the underlying turf every time you crowbar up the tile, leading to a property where it resets the icon_state of the desert or road underneath.

In the former it sets it to something random and can even spawn grasses or, if random wildplants are added, could force them to spawn/despawn.

In the latter it resets it to icon_state outermiddle.

This is caused because adding tiles literally transforms the turf into a completely different turf.

It is less of an obnoxious issue than a random purple-and-pink tile that crops every time you put floor tile down but still A Problem.