Fixes
Attempts (unsuccessfully) to fix runtimes where blood is called to be added to a weapon after a target has been gibbed by a melee weapon by moving damage to the end of various procs.

Changes requirement for being being deleted to health*4 or *1.9 for toxin, due to toxin itself being limited to 200.