Grammar and descriptions
Adds updating descriptions to the doors and such so you can shift-click on them to see if they have a lock and what number it is.

Still need to add an overlay update so that the lock visibly gets added to the door, a sprite for that, and maybe a better lock sprite.