Fixes crosses
Makes it so unbuckling from crosses actually works.

Makes it so they don't arbitrarily damage you just for being strung up on one.

Makes it so you aren't rotated upside-down when you get buckled to a cross (Really?)

Plays the correct noises now.

Does the correct animations now.

Very minorly reduces damage from being on a kitchen spike so you don't just resist free and then instantly die of bloodloss.

There's probably a way to rewrite this chunk of code so there's less copypasting, but fuck it.

tl;dr the entire cross code was busted and not being used and it was defaulting to kitchenspike code