Update BoxStation.dmm
Description
Add a new building (gym) to next the Market

Motivation and Context
I believe there is a need for more structures around the city/Den, as there is a lot of unused space on the map currently

How Has This Been Tested?
tested, by running it trough ingame

Screenshots.
Bos holotag