Slight map audit
Removes the deathclaw nest from the lastway
Adds a radroach nest further south on top of the loot
Adds a smarter/stupider loot drop (it suffers the issue of grabbing ANY type of seed, so you might see a revolver plant or something; it'll need a bit of working to do what it's supposed to).
Changes the feedback when joining as an observer.
Adds a vault-tek VERIFIED EXPERIMENT PAPER to the overseer's desk that picks from a small list.