Changes job hours
Along with stripping out a few jobs (namely all the prime stuff, what the hell) this also lowers the required hours playtime for a lot of things.

It's been objectively demonstrated that it doesn't stop terrible players, it only acts as a temporary filter, and we've been having weird behavior where people pick a role then afk as it to open up higher slots.