Taking away benches
I might go further and strip out the ammobench and constructionlathe in parump and replace it with the autolathe again.

The whole idea was to make it so that ammunition and gear eventually becomes scarce, despite initial up-front content for a newspawn character (so they don't end up utterly and arbitrarily screwed) and that the resource to spawn more of it becomes an obsession for BoS or whatever;

I can't say it makes sense for the BoS to steal an ammo bench but I *can* say it makes sense for them to take a magical wish machine that turns infinite amounts of material into an endless supply of grenade casings, bullets, gear, and other ammunition that is useful for waging mass war.

This is why constructionlathe/autolathe separation was made in the first place.